% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celldiveUtils-package.R
\docType{package}
\name{celldiveUtils-package}
\alias{celldiveUtils}
\alias{celldiveUtils-package}
\title{celldiveUtils: Helper Functions to Analyse Segmented and Quantified CellDIVE Data}
\description{
Set of helper and auxillary function to analyse and perform clustering of previously segmented and quantified CellDIVE images. This package should only be use within the DIVEMAP pipeline.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
